+++ packages/tiny-canvas/src/index.js
@@ -5,4 +5,5 @@ export { default as Block } from './Block';
 export { default as Frame } from './Frame';
 export { default as Note } from './Note';
 export { default as Mark } from './Mark';
+export { default as Link } from './Link';
 export { useResetCanvas } from './useResetCanvas';